Moved common to convert_utils.py
